Extract actions
