+++ apps/spacerage/src/scenes/gameplay/Debris.tsx
@@ -1,12 +1,16 @@
-import { GroupProps } from "@react-three/fiber"
+import { GroupProps, Object3DProps } from "@react-three/fiber"
 import { Composable, Modules } from "material-composer-r3f"
-import { createContext } from "react"
+import { between, upTo } from "randomish"
+import { createContext, useContext } from "react"
 import { $, Input, InstanceID, Mul, OneMinus, Vec3 } from "shader-composer"
 import { Random } from "shader-composer-toybox"
-import { Particles, useParticleLifetime } from "vfx-composer-r3f"
+import { Vector3 } from "three"
+import { Emitter, Particles, useParticleLifetime } from "vfx-composer-r3f"
 import { ECS } from "./state"
 
 export const DebrisContext = createContext<{ particles: any }>(null!)
+
+const tmpVec3 = new Vector3()
 
 export const Debris = (props: GroupProps) => {
   const particles = useParticleLifetime()
@@ -39,5 +43,22 @@ export const Debris = (props: GroupProps) => {
         </DebrisContext.Provider>
       </Particles>
     </group>
+  )
+}
+
+export const DebrisEmitter = (props: Object3DProps) => {
+  const { particles } = useContext(DebrisContext)
+
+  return (
+    <Emitter
+      {...props}
+      rate={Infinity}
+      limit={between(2, 5)}
+      setup={({ position, scale }) => {
+        scale.setScalar(between(0.5, 2))
+        particles.setLifetime(between(0.5, 1.5), upTo(0.1))
+        position.add(tmpVec3.randomDirection())
+      }}
+    />
   )
 }
